polymorphism evaluators
